+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/configuration/Configuration.java
@@ -966,12 +966,7 @@ public final class Configuration {
     
     @ConfigurationDefaultBoolean(true)
     @ConfigurationCategory("ui")
-    public static ConfigurationItem<Boolean> displayAs3PCodePanel = null;
-    
-    @ConfigurationDefaultBoolean(true)
-    @ConfigurationName("warning.replace.morphshape")
-    @ConfigurationCategory("ui")
-    public static ConfigurationItem<Boolean> warningReplaceMorphShape = null;
+    public static ConfigurationItem<Boolean> displayAs3PCodePanel = null;       
         
     private enum OSId {
         WINDOWS, OSX, UNIX