+++ CHANGELOG.md
@@ -13,7 +13,7 @@ All notable changes to this project will be documented in this file.
 - [#2105] GFX - Basic tag info
 - Context menu items to create new tags (shape, morphshape, sprite, image, movie, sound, binaryData) from files
   and using font embed dialog for fonts
-- Replacing morphshapes (currently only same shape for start/end)
+- Replacing morphshapes
 
 ### Fixed
 - [#1306], [#1768] Maximizing window on other than main monitor
@@ -24,6 +24,7 @@ All notable changes to this project will be documented in this file.
 - [#2104] Empty texts import
 - Centered start playing triangle (Playing on demand)
 - miterLimitFactor is FIXED8 value in MORPHLINESTYLE2
+- Display of morphshape end shape to be exactly at 65535 ratio
 
 ### Changed
 - Basic tag info panel always visible even when nothing to display (to avoid flickering)